add 404 page change search bar in path seach
