+++ src/pages/index.tsx
@@ -4,3 +4,4 @@ export * from './Login'
 export * from './Profile'
 export * from './Register'
 export * from './Success'
+export * from './NotFound'
