+++ src/style/theme.less
@@ -53,3 +53,11 @@ code {
 .ant-form-item-has-error :not(.ant-input-disabled).ant-input:hover{
   box-shadow: none;
 }
+
+.centered {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  /* bring your own prefixes */
+  transform: translate(-50%, -50%);
+}
